requirements.txt
streamlit
pandas
numpy
plotly
seaborn
matplotlib
scikit-learn
openpyxl
scipy
google-generativeai
python-dotenv 